rename package to Performant UI
