+++ packages/performant-ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tailwind-ui",
+  "name": "performant-ui",
   "version": "1.0.0",
   "main": "./dist/index.cjs.js",
   "module": "./dist/index.es.js",
