+++ packages/storybook/src/performant-ui.css
@@ -1,0 +1,5 @@
+@import 'tailwindcss';
+@import '../../performant-ui/tailwind.css';
+@source '../../performant-ui/';
+
+@custom-variant dark (&:where(.dark, .dark *));
