--- packages/storybook/src/tailwind-ui.css
@@ -1,5 +1,0 @@
-@import 'tailwindcss';
-@import '../../tailwind-ui/tailwind.css';
-@source '../../tailwind-ui/';
-
-@custom-variant dark (&:where(.dark, .dark *));
